refactor: separate `test/cpp.ts` into files under `test/cpp` dir
