+++ src/test/cpp/compile.ts
@@ -3,7 +3,7 @@ import { mkdirSync } from "node:fs";
 import path from "node:path";
 
 /**
- * Compile a C++ test file using Clang.
+ * Compiles a C++ test file using Clang.
  *
  * @param testFile - The path of the C++ test file to compile.
  * @returns A path to the compiled test executable.
@@ -18,13 +18,4 @@ export function compileCppTest(testFile: string): string {
   });
 
   return testExec;
-}
-
-/**
- * Run a C++ test executable.
- *
- * @param testExec - The path of the C++ test executable to run.
- */
-export function runCppTest(testExec: string): void {
-  execSync(testExec, { stdio: "inherit" });
 }
