+++ src/test/cpp/index.ts
@@ -1,0 +1,2 @@
+export { compileCppTest } from "./compile.js";
+export { runCppTest } from "./run.js";
