+++ src/test/cpp/run.ts
@@ -1,0 +1,10 @@
+import { execSync } from "node:child_process";
+
+/**
+ * Runs a C++ test executable.
+ *
+ * @param testExec - The path of the C++ test executable to run.
+ */
+export function runCppTest(testExec: string): void {
+  execSync(testExec, { stdio: "inherit" });
+}
